+++ OUDS/Core/Themes/Sosh/Sources/Values/SemanticTokens/SoshTheme+ColorMultipleSemanticTokens.swift
@@ -21,8 +21,6 @@ import OUDSTokensSemantic
 // swiftlint:disable line_length
 
 /// Defines provider objects for `ColorSemanticTokens` so as to pack them as light and dark mode colors.
-/// These values can be overriden inside `SoshThemeColorSemanticTokensProvider` subclasses (in extensions or not, in the same module or not) thanks to the `@objc open` combination.
-/// Some tokens do not have values assigned in the design system, and must be overriden.
 /// Helps to expose color semantic tokens with two values to use depending to the color scheme (*Figma* cannot manage such tokens and generate them).
 extension SoshThemeColorSemanticTokensProvider: ColorMultipleSemanticTokens {
 
@@ -118,7 +116,7 @@ extension SoshThemeColorSemanticTokensProvider: ColorMultipleSemanticTokens {
 
     @objc public final var borderOnBrandTertiary: MultipleColorSemanticTokens { MultipleColorSemanticTokens(light: borderOnBrandTertiaryLight, dark: borderOnBrandTertiaryDark) }
 
-    @objc public final var borderMinmal: MultipleColorSemanticTokens { MultipleColorSemanticTokens(light: borderMinimalLight, dark: borderMinimalDark) }
+    @objc public final var borderMinimal: MultipleColorSemanticTokens { MultipleColorSemanticTokens(light: borderMinimalLight, dark: borderMinimalDark) }
 
     // MARK: - Color - Content