+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- `color` semantic tokens (tokens libraries Core v1.9, System v2.2)
 - `icon` component tokens (tokens libraries Core v1.9, System v2.2)
 - `control item` component tokens (tokens libraries Core v1.9, System v2.2) 
 - Update `text input` to display the label on two lines and in the middle when no placeholder and one line at top otherwise (Orange-OpenSource/ouds-ios#1086)